Move SpacingAroundAngleBracketsRule from Experimental to Standard ruleset
